+++ NEWS.md
@@ -13,6 +13,7 @@
 * Remove integrity checks for concatenating relations or databases together, for performance: since the only change is some renaming, concatenating valid relations/databases shouldn't result in any violations anyway.
 * Added support for skipping simple keys, and bijections if `skip_bijections = TRUE` in `discover`, for all methods, rather than just DFD.
 * Added support for skipping fixed attributes, simple keys, and bijections if `skip_bijections = TRUE` in `discover`, for `discover_keys`.
+* Added `autokey` to add keys to a given data frame. This is to `discover_keys` as `autodb` is to `discover`.
 
 # autodb 3.2.4